+++ sdk/appcenter-distribute/src/main/java/com/microsoft/appcenter/distribute/download/manager/DownloadManagerUpdateTask.java
@@ -38,18 +38,19 @@ class DownloadManagerUpdateTask extends AsyncTask<Void, Void, Void> {
         try {
             Cursor cursor = downloadManager.query(new DownloadManager.Query().setFilterById(downloadId));
             if (cursor == null) {
-                throw new NoSuchElementException();
+                throw new NoSuchElementException("Cannot find download with id=" + downloadId);
             }
             try {
                 if (!cursor.moveToFirst()) {
-                    throw new NoSuchElementException();
+                    throw new NoSuchElementException("Cannot find download with id=" + downloadId);
                 }
                 if (isCancelled()) {
                     return null;
                 }
                 int status = cursor.getInt(cursor.getColumnIndexOrThrow(DownloadManager.COLUMN_STATUS));
                 if (status == DownloadManager.STATUS_FAILED) {
-                    throw new IllegalStateException();
+                    int reason = cursor.getInt(cursor.getColumnIndexOrThrow(DownloadManager.COLUMN_REASON));
+                    throw new IllegalStateException("The download has failed with reason code: " + reason);
                 }
                 if (status != DownloadManager.STATUS_SUCCESSFUL) {
                     mDownloader.onDownloadProgress(cursor);